+++ app/lib/ui/components/checkbox/checkbox_item_demo_screen.dart
@@ -89,7 +89,7 @@ class _BodyState extends State<_Body> {
       widget: Column(
         children: [
           _CheckboxItemDemo(indeterminate: widget.indeterminate),
-          SizedBox(height: themeController.currentTheme.spaceTokens.fixedTall),
+          SizedBox(height: themeController.currentTheme.spaceScheme(context).fixedTall),
           Code(
             code: ControlItemCodeGenerator.updateCode(context, widget.indeterminate, ControlItemType.checkbox),
           ),