+++ app/lib/ui/tokens/color/color_screen.dart
@@ -40,7 +40,7 @@ class ColorScreen extends StatelessWidget {
               fit: BoxFit.fitWidth,
             ),
             Padding(
-              padding: EdgeInsetsDirectional.all(currentTheme.spaceTokens.paddingInlineTall),
+              padding: EdgeInsetsDirectional.all(currentTheme.spaceScheme(context).paddingInlineTall),
               child: Column(
                 children: [
                   Text(
@@ -58,15 +58,15 @@ class ColorScreen extends StatelessWidget {
               shrinkWrap: true,
               physics: NeverScrollableScrollPhysics(),
               padding: EdgeInsetsDirectional.only(
-                top: currentTheme.spaceTokens.paddingInlineTall,
-                bottom: currentTheme.spaceTokens.paddingInlineTall,
+                top: currentTheme.spaceScheme(context).paddingInlineTall,
+                bottom: currentTheme.spaceScheme(context).paddingInlineTall,
               ),
               children: [
                 for (var entry in tokenGroups.entries) ...[
                   Padding(
                     padding: EdgeInsetsDirectional.symmetric(
-                      vertical: currentTheme.spaceTokens.rowGapTall,
-                      horizontal: currentTheme.spaceTokens.paddingInlineTall,
+                      vertical: currentTheme.spaceScheme(context).rowGapTall,
+                      horizontal: currentTheme.spaceScheme(context).paddingInlineTall,
                     ),
                     child: Semantics(
                       header: true,
@@ -84,7 +84,7 @@ class ColorScreen extends StatelessWidget {
                   ...entry.value.map(
                     (item) => Padding(
                       padding: EdgeInsetsDirectional.symmetric(
-                        horizontal: currentTheme.spaceTokens.paddingInlineTall,
+                        horizontal: currentTheme.spaceScheme(context).paddingInlineTall,
                       ),
                       child: ColorWidget(colorTokenItem: item),
                     ),
@@ -110,7 +110,7 @@ class ColorWidget extends StatelessWidget {
     final currentTheme = themeController.currentTheme;
 
     return Padding(
-      padding: EdgeInsetsDirectional.symmetric(vertical: currentTheme.spaceTokens.rowGapShort),
+      padding: EdgeInsetsDirectional.symmetric(vertical: currentTheme.spaceScheme(context).rowGapShort),
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
@@ -122,7 +122,7 @@ class ColorWidget extends StatelessWidget {
               border: Border.all(color: currentTheme.colorScheme(context).actionEnabled),
             ),
           ),
-          SizedBox(width: currentTheme.spaceTokens.paddingInlineTall),
+          SizedBox(width: currentTheme.spaceScheme(context).paddingInlineTall),
           Expanded(
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
@@ -136,7 +136,7 @@ class ColorWidget extends StatelessWidget {
                     color: currentTheme.colorScheme(context).contentDefault,
                   ),
                 ),
-                SizedBox(height: currentTheme.spaceTokens.rowGapNone),
+                SizedBox(height: currentTheme.spaceScheme(context).rowGapNone),
                 Text(
                   colorTokenItem.colorToHex(colorTokenItem.value),
                   style: TextStyle(