+++ ouds_theme_white_label/lib/white_label_theme.dart
@@ -11,9 +11,11 @@
  */
 
 import 'package:flutter/material.dart';
-import 'package:ouds_theme_contract/ouds_color_scheme.dart';
 import 'package:ouds_theme_contract/ouds_theme_contract.dart';
 import 'package:ouds_theme_contract/ouds_tokens_provider.dart';
+import 'package:ouds_theme_contract/theme/scheme/color/ouds_color_scheme.dart';
+import 'package:ouds_theme_contract/theme/scheme/responsive/ouds_grid_scheme.dart';
+import 'package:ouds_theme_contract/theme/scheme/responsive/ouds_space_scheme.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_button_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_checkbox_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_components_tokens.dart';
@@ -83,13 +85,13 @@ class WhiteLabelTheme implements OudsThemeContract {
   OudsElevationSemanticTokens get elevationTokens => WhiteLabelElevationSemanticTokens();
 
   @override
-  OudsSpaceSemanticTokens get spaceTokens => WhiteLabelSpaceSemanticTokens();
+  OudsSpaceSemanticTokens get spaceSemanticTokens => WhiteLabelSpaceSemanticTokens();
 
   @override
   OudsSizeSemanticTokens get sizeTokens => WhiteLabelSizeSemanticTokens();
 
   @override
-  OudsGridSemanticTokens get gridTokens => WhiteLabelGridSemanticTokens();
+  OudsGridSemanticTokens get gridSemanticTokens => WhiteLabelGridSemanticTokens();
 
   @override
   OudsFontSemanticTokens get fontTokens => WhiteLabelFontSemanticTokens();
@@ -105,19 +107,15 @@ class WhiteLabelTheme implements OudsThemeContract {
     return OudsColorScheme.fromContext(context: context, colorTokens: colorSemanticTokens);
   }
 
-  /*
   @override
-  OudsProvidersTokens get providersTokens => OudsProvidersTokens(
-      colorScheme: colorsScheme,
-      opacityTokens: opacityTokens,
-      borderTokens: borderTokens,
-      elevationTokens: elevationTokens,
-      spaceTokens: spaceTokens,
-      sizeTokens: sizeTokens,
-      gridTokens: gridTokens,
-      fontTokens: fontTokens);
+  OudsSpaceScheme spaceScheme(BuildContext context) {
+    return OudsSpaceScheme.fromContext(context: context, spaceTokens: spaceSemanticTokens);
+  }
 
-   */
+  @override
+  OudsGridScheme gridScheme(BuildContext context) {
+    return OudsGridScheme.fromContext(context: context, gridTokens: gridSemanticTokens);
+  }
 
   @override
   OudsProvidersTokens providersTokens(BuildContext context) {
@@ -126,9 +124,9 @@ class WhiteLabelTheme implements OudsThemeContract {
       opacityTokens: opacityTokens,
       borderTokens: borderTokens,
       elevationTokens: elevationTokens,
-      spaceTokens: spaceTokens,
+      spaceTokens: spaceSemanticTokens,
+      spaceScheme: spaceScheme(context),
       sizeTokens: sizeTokens,
-      gridTokens: gridTokens,
       fontTokens: fontTokens,
     );
   }